+++ packages/analytics-js-plugins/src/nativeDestinationQueue/index.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 /* eslint-disable no-param-reassign */
 import { ApplicationState } from '@rudderstack/analytics-js-common/types/ApplicationState';
-import { IPluginsManager } from '@rudderstack/analytics-js-common/types/PluginsManager';
+import { IPluginsManager, PluginName } from '@rudderstack/analytics-js-common/types/PluginsManager';
 import { IStoreManager } from '@rudderstack/analytics-js-common/types/Store';
 import { IErrorHandler } from '@rudderstack/analytics-js-common/types/ErrorHandler';
 import { ILogger } from '@rudderstack/analytics-js-common/types/Logger';
@@ -9,16 +9,16 @@ import { QueueOpts } from '@rudderstack/analytics-js-common/types/LoadOptions';
 import { RudderEvent } from '@rudderstack/analytics-js-common/types/Event';
 import { Destination } from '@rudderstack/analytics-js-common/types/Destination';
 import { ExtensionPlugin } from '@rudderstack/analytics-js-common/types/PluginEngine';
-import { MEMORY_STORAGE } from '@rudderstack/analytics-js-common/constants/storages';
 import { clone } from 'ramda';
+import { storages } from '../shared-chunks/common';
 import { DoneCallback, IQueue } from '../types/plugins';
 import { RetryQueue } from '../utilities/retryQueue/RetryQueue';
 import { getNormalizedQueueOptions, isEventDenyListed, sendEventToDestination } from './utilities';
 import { NATIVE_DESTINATION_QUEUE_PLUGIN, QUEUE_NAME } from './constants';
-import { DESTINATION_EVENT_FILTERING_WARNING } from '../utilities/logMessages';
-import { filterDestinations } from '../utilities/destination';
+import { DESTINATION_EVENT_FILTERING_WARNING } from './logMessages';
+import { destinationUtils } from '../shared-chunks/deviceModeDestinations';
 
-const pluginName = 'NativeDestinationQueue';
+const pluginName: PluginName = 'NativeDestinationQueue';
 
 const NativeDestinationQueue = (): ExtensionPlugin => ({
   name: pluginName,
@@ -54,7 +54,7 @@ const NativeDestinationQueue = (): ExtensionPlugin => ({
         `${QUEUE_NAME}_${writeKey}`,
         finalQOpts,
         (rudderEvent: RudderEvent, done: DoneCallback) => {
-          const destinationsToSend = filterDestinations(
+          const destinationsToSend = destinationUtils.filterDestinations(
             rudderEvent.integrations,
             state.nativeDestinations.initializedDestinations.value,
           );
@@ -103,7 +103,7 @@ const NativeDestinationQueue = (): ExtensionPlugin => ({
           done(null);
         },
         storeManager,
-        MEMORY_STORAGE,
+        storages.MEMORY_STORAGE,
       );
 
       // TODO: This seems to not work as expected. Need to investigate