+++ packages/analytics-js/src/components/pluginsManager/PluginsManager.ts
@@ -9,6 +9,7 @@ import { IErrorHandler } from '@rudderstack/analytics-js-common/types/ErrorHandl
 import { ILogger } from '@rudderstack/analytics-js-common/types/Logger';
 import { Nullable } from '@rudderstack/analytics-js-common/types/Nullable';
 import { PLUGINS_MANAGER } from '@rudderstack/analytics-js-common/constants/loggerContexts';
+import { isFunction } from '@rudderstack/analytics-js-common/utilities/checks';
 import { setExposedGlobal } from '../utilities/globals';
 import { state } from '../../state';
 import {
@@ -17,7 +18,7 @@ import {
   StorageEncryptionVersionsToPluginNameMap,
 } from '../configManager/constants';
 import { UNSUPPORTED_BEACON_API_WARNING } from '../../constants/logMessages';
-import { remotePluginNames } from './pluginNames';
+import { pluginNamesList } from './pluginNames';
 import {
   getMandatoryPluginsMap,
   pluginsInventory,
@@ -112,7 +113,7 @@ class PluginsManager implements IPluginsManager {
       );
     }
 
-    // dataplane events delivery plugins
+    // Cloud mode (dataplane) events delivery plugins
     if (state.loadOptions.value.useBeacon === true && state.capabilities.isBeaconAvailable.value) {
       pluginsToLoadFromConfig = pluginsToLoadFromConfig.filter(
         pluginName => pluginName !== 'XhrQueue',
@@ -125,6 +126,14 @@ class PluginsManager implements IPluginsManager {
       pluginsToLoadFromConfig = pluginsToLoadFromConfig.filter(
         pluginName => pluginName !== 'BeaconQueue',
       );
+    }
+
+    // Enforce default cloud mode event delivery queue plugin is none exists
+    if (
+      !pluginsToLoadFromConfig.includes('XhrQueue') &&
+      !pluginsToLoadFromConfig.includes('BeaconQueue')
+    ) {
+      pluginsToLoadFromConfig.push('XhrQueue');
     }
 
     // Device mode destinations related plugins
@@ -180,7 +189,7 @@ class PluginsManager implements IPluginsManager {
   setActivePlugins() {
     const pluginsToLoad = this.getPluginsToLoadBasedOnConfig();
     // Merging available mandatory and optional plugin name list
-    const availablePlugins = [...Object.keys(pluginsInventory), ...remotePluginNames];
+    const availablePlugins = [...Object.keys(pluginsInventory), ...pluginNamesList];
     const activePlugins: PluginName[] = [];
     const failedPlugins: string[] = [];
 
@@ -216,7 +225,10 @@ class PluginsManager implements IPluginsManager {
    */
   registerLocalPlugins() {
     Object.values(pluginsInventory).forEach(localPlugin => {
-      if (state.plugins.activePlugins.value.includes(localPlugin().name)) {
+      if (
+        isFunction(localPlugin) &&
+        state.plugins.activePlugins.value.includes(localPlugin().name)
+      ) {
         this.register([localPlugin()]);
       }
     });
@@ -232,7 +244,7 @@ class PluginsManager implements IPluginsManager {
 
     Promise.all(
       Object.keys(remotePluginsList).map(async remotePluginKey => {
-        await remotePluginsList[remotePluginKey]()
+        await remotePluginsList[remotePluginKey as PluginName]()
           .then((remotePluginModule: any) => this.register([remotePluginModule.default()]))
           .catch(err => {
             // TODO: add retry here if dynamic import fails