+++ packages/analytics-js/src/components/pluginsManager/defaultPluginsList.ts
@@ -3,22 +3,21 @@ import { PluginName } from '@rudderstack/analytics-js-common/types/PluginsManage
 /**
  * Plugins to be loaded in the plugins loadOption is not defined
  */
-// TODO: remove from this list the ones that should be loaded only if specific load options exist
 const defaultOptionalPluginsList: PluginName[] = [
+  'BeaconQueue',
   'Bugsnag',
   'DeviceModeDestinations',
+  'DeviceModeTransformation',
   'ErrorReporting',
   'ExternalAnonymousId',
   'GoogleLinker',
+  'KetchConsentManager',
   'NativeDestinationQueue',
+  'OneTrustConsentManager',
   'StorageEncryption',
   'StorageEncryptionLegacy',
   'StorageMigrator',
   'XhrQueue',
-  'OneTrustConsentManager',
-  'KetchConsentManager',
-  'BeaconQueue',
-  'DeviceModeTransformation',
 ];
 
 export { defaultOptionalPluginsList };