+++ packages/analytics-js/src/components/pluginsManager/types.ts
@@ -1,3 +1,4 @@
 import { ExtensionPlugin } from '@rudderstack/analytics-js-common/types/PluginEngine';
+import { PluginName } from '@rudderstack/analytics-js-common/types/PluginsManager';
 
-export type PluginMap<T = ExtensionPlugin> = Record<string, () => T>;
+export type PluginMap<T = ExtensionPlugin> = Record<PluginName, () => T>;